requirements.txt
anthropic
langchain
langchain-community
langchain-text-splitters
pymupdf
chromadb
duckdb
pydantic
streamlit
plotly
pandas
matplotlib
sentence-transformers
python-dotenv
pytest
